Fixed Filter as a component
